+++ changelog.md
@@ -5,8 +5,7 @@
 ## 2.3.0
 
 ###### 2024-01-10
-Added ability to change date added book in book status slide menu
-- [New] 
+- [New] Added ability to change date added book in book status slide menu
 
 ## 2.2.3